requirements.txt
gradio==5.9.1
tf-keras==2.18.0
transformers==4.47.1
torch==2.5.1
Pillow==11.0.0